코드 수정
이용권 구매, 대여소,, 메인화면 코드 수정 ver 2

+++ src/main/java/com/database4/dto/ReturnGetMapLocationDto.java
@@ -3,17 +3,19 @@ package com.database4.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class ReturnGetMapLocationDto {
     private int bikeCount;
-    private float latitude;
-    private float longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
 
     public ReturnGetMapLocationDto() {
     }
 
-    public ReturnGetMapLocationDto(float latitude, float longitude, int bikeCount) {
+    public ReturnGetMapLocationDto(BigDecimal latitude, BigDecimal longitude, int bikeCount) {
         this.latitude = latitude;
         this.longitude = longitude;
         this.bikeCount = bikeCount;
